+++ docs/example_scraper.rb
@@ -1,6 +1,8 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
+Bundler.require
+
 $LOAD_PATH << "./lib"
 
 require "scraper_utils"
@@ -57,7 +59,7 @@ class Scraper
     unless exceptions.empty?
       puts "\n***************************************************"
       puts "Now retrying authorities which earlier had failures"
-      puts exceptions.keys.join(", ").to_s
+      puts exceptions.keys.join(", ")
       puts "***************************************************"
       ENV['DEBUG'] ||= '1'
 
@@ -80,8 +82,21 @@ end
 if __FILE__ == $PROGRAM_NAME
   # Default to list of authorities we can't or won't fix in code, explain why
   # some: url-for-issue Summary Reason
-  # councils : url-for-issue Summary Reason
+  # councils: url-for-issue Summary Reason
 
-  ENV["MORPH_EXPECT_BAD"] ||= "some,councils"
+  if ENV['MORPH_EXPECT_BAD'].nil?
+    default_expect_bad = {
+    }
+    puts 'Default EXPECT_BAD:', default_expect_bad.to_yaml if default_expect_bad.any?
+
+    ENV["MORPH_EXPECT_BAD"] = default_expect_bad.keys.join(',')
+  end
   Scraper.run(Scraper.selected_authorities)
+
+  # Dump database for morph-cli
+  if File.exist?("tmp/dump-data-sqlite")
+    puts "-- dump of data.sqlite --"
+    system "sqlite3 data.sqlite .dump"
+    puts "-- end of dump --"
+  end
 end